Add support for GOT
This commit add basic support for GOT accesses. During relocation
processing (linking) a relocation may request GOT entry for its symbol
by

    gotentry := AcRelocationRequestGOTEntry query: self

The response (`gotentry` in the above example) is an association from
GOT entry symbol to its address (that is, address of the entry). Actual
allocation of GOT entries is the responsibility of the linker.