DSL: refactor assembling (encoding) of instructions
This commit refactors assembling (encoding) of DSL operands by
delegating the job to individual operand format chunks. This makes the
code more flexible and extensible.

While at it, this commit also introduces checks if DSL operands are what
is expected and throw (new) `AcDSLInvalidOperandError` if not. This
makes the code more robust w.r.t. misuse (aka it should be impossible
to construct invalid structure) and makes the code ready for ISAs where
instruction cannot be determined solely by mnemonic (AMD64, AArch64 for
instance).

The downside of these checks is that they do not really work with POWER
because of a way the PDL is done. It uses `%imm` instead of register
map (like %[gpr] or something), exploiting the fact that in POWER
assembly language, registers are identified by their number. So to make
things work, this commit includes temporary hack to skip the check
on POWER. Temporary till we get a better quality POWER model.