DSL: generalize assembling (encoding) logic
This commit generalizes DSL assembling (encoding) to support ISAs where
the instruction encoded also depends on operands (and not solely in
mnemonic as current code assumes).

This is implemented leveraging previous improvements that raise an
exception if operands are invalid (do not match the syntax/format).
In that case, continue trying other other instructions.

While at it, add comment a raise `AcDSLInvalidInstructionError` if
there's no such instruction (instead of generic `Error`).

Note that this code can be sped up by maintaining a map from mnemonic
to list of instructions with that mnemonic. This is left as future
work.